requirements.txt
numpy
pandas
scikit-learn
streamlit
matplotlib
pillow
diffusers
replicate
